+++ cf/functions.py
@@ -157,6 +157,7 @@ def configuration(
     chunksize=None,
     log_level=None,
     display_data=None,
+    persist_data=None,
     regrid_logging=None,
     relaxed_identities=None,
     bounds_combination_mode=None,
@@ -179,6 +180,7 @@ def configuration(
     * `chunksize`
     * `log_level`
     * `display_data`
+    * `persist_data`
     * `regrid_logging`
     * `relaxed_identities`
     * `bounds_combination_mode`
@@ -203,9 +205,10 @@ def configuration(
 
     .. seealso:: `atol`, `rtol`, `tempdir`, `chunksize`,
                  `total_memory`, `log_level`, `display_data`,
-                 `regrid_logging`, `relaxed_identities`,
-                 `bounds_combination_mode`, `active_storage`,
-                 `active_storage_url`, `active_storage_max_requests`
+                 `persist_data`, `regrid_logging`,
+                 `relaxed_identities`, `bounds_combination_mode`,
+                 `active_storage`, `active_storage_url`,
+                 `active_storage_max_requests`
 
     :Parameters:
 
@@ -247,11 +250,17 @@ def configuration(
             * ``'DETAIL'`` (``3``);
             * ``'DEBUG'`` (``-1``).
 
-        display_data `bool` or `Constant`, optional
+        display_data: `bool` or `Constant`, optional
             The new display data option. The default is to not change
             the current behaviour.
 
             .. versionadded:: 3.19.0
+
+        persist_data: `bool` or `Constant`, optional
+            The new persist data option. The default is to not change
+            the current behaviour.
+
+            .. versionadded:: NEXTVERSION
 
         regrid_logging: `bool` or `Constant`, optional
             The new value (either True to enable logging or False to
@@ -312,6 +321,7 @@ def configuration(
      'bounds_combination_mode': 'AND',
      'chunksize': 82873466.88000001,
      'display_data': True,
+     'persist_data': False,
      'active_storage': False,
      'active_storage_url': None,
      'active_storage_max_requests': 100}
@@ -330,6 +340,7 @@ def configuration(
      'bounds_combination_mode': 'AND',
      'chunksize': 75000000.0,
      'display_data': True,
+     'persist_data': False,
      'active_storage': False,
      'active_storage_url': None,
      'active_storage_max_requests': 100}
@@ -358,6 +369,7 @@ def configuration(
      'bounds_combination_mode': 'AND',
      'chunksize': 75000000.0,
      'display_data': True,
+     'persist_data': False,
      'active_storage': False,
      'active_storage_url': None}
     >>> with cf.configuration(atol=9, rtol=10):
@@ -372,6 +384,7 @@ def configuration(
      'bounds_combination_mode': 'AND',
      'chunksize': 75000000.0,
      'display_data': True,
+     'persist_data': False,
      'active_storage': False,
      'active_storage_url': None,
      'active_storage_max_requests': 100}
@@ -385,6 +398,7 @@ def configuration(
      'bounds_combination_mode': 'AND',
      'chunksize': 75000000.0,
      'display_data': True,
+     'persist_data': False,
      'active_storage': False,
      'active_storage_url': None,
      'active_storage_max_requests': 100}
@@ -416,6 +430,7 @@ def configuration(
         new_chunksize=chunksize,
         new_log_level=log_level,
         new_display_data=display_data,
+        new_persist_data=persist_data,
         new_regrid_logging=regrid_logging,
         new_relaxed_identities=relaxed_identities,
         bounds_combination_mode=bounds_combination_mode,
@@ -460,6 +475,7 @@ def _configuration(_Configuration, **kwargs):
         "new_chunksize": chunksize,
         "new_log_level": log_level,
         "new_display_data": display_data,
+        "new_persist_data": persist_data,
         "new_regrid_logging": regrid_logging,
         "new_relaxed_identities": relaxed_identities,
         "bounds_combination_mode": bounds_combination_mode,
@@ -587,6 +603,10 @@ class log_level(ConstantAccess, cfdm.log_level):
 
 
 class display_data(ConstantAccess, cfdm.display_data):
+    pass
+
+
+class persist_data(ConstantAccess, cfdm.persist_data):
     pass
 
 
@@ -3183,60 +3203,6 @@ def environment(display=True, paths=True):
             If *display* is True then the description of the
             environment is printed and `None` is returned. Otherwise
             the description is returned as a string.
-
-    **Examples**
-
-    >>> cf.environment()
-    Platform: Linux-6.8.0-60-generic-x86_64-with-glibc2.39
-    Python: 3.12.8 /home/miniconda3/bin/python
-    packaging: 24.2 /home/miniconda3/lib/python3.12/site-packages/packaging/__init__.py
-    numpy: 2.2.6 /home/miniconda3/lib/python3.12/site-packages/numpy/__init__.py
-    cfdm.core: 1.12.2.0 /home/miniconda3/lib/python3.12/site-packages/cfdm/cfdm/core/__init__.py
-    udunits2 library: libudunits2.so.0
-    HDF5 library: 1.14.2
-    netcdf library: 4.9.4-development
-    netCDF4: 1.7.2 /home/miniconda3/lib/python3.12/site-packages/netCDF4/__init__.py
-    h5netcdf: 1.3.0 /home/miniconda3/lib/python3.12/site-packages/h5netcdf/__init__.py
-    h5py: 3.12.1 /home/miniconda3/lib/python3.12/site-packages/h5py/__init__.py
-    zarr: 3.1.3 /home/miniconda3/lib/python3.12/site-packages/zarr/__init__.py
-    s3fs: 2024.12.0 /home/miniconda3/lib/python3.12/site-packages/s3fs/__init__.py
-    scipy: 1.15.1 /home/miniconda3/lib/python3.12/site-packages/scipy/__init__.py
-    dask: 2025.5.1 /home/miniconda3/lib/python3.12/site-packages/dask/__init__.py
-    distributed: 2025.5.1 /home/miniconda3/lib/python3.12/site-packages/distributed/__init__.py
-    cftime: 1.6.4.post1 /home/miniconda3/lib/python3.12/site-packages/cftime/__init__.py
-    cfunits: 3.3.7 /home/miniconda3/lib/python3.12/site-packages/cfunits/__init__.py
-    cfdm: 1.12.2.0 /home/miniconda3/lib/python3.12/site-packages/cfdm/__init__.py
-    esmpy/ESMF: 8.7.0 /home/miniconda3/lib/python3.12/site-packages/esmpy/__init__.py
-    psutil: 6.1.1 /home/miniconda3/lib/python3.12/site-packages/psutil/__init__.py
-    matplotlib: 3.10.0 /home/miniconda3/lib/python3.12/site-packages/matplotlib/__init__.py
-    cfplot: 3.4.0 /home/miniconda3/lib/python3.12/site-packages/cfplot/__init__.py
-    cf: 3.18.0 /home/miniconda3/lib/python3.12/site-packages/cf/__init__.py
-
-    >>> cf.environment(paths=False)
-    Platform: Linux-6.8.0-60-generic-x86_64-with-glibc2.39
-    Python: 3.12.8
-    packaging: 24.2
-    numpy: 2.2.6
-    cfdm.core: 1.12.2.0
-    udunits2 library: libudunits2.so.0
-    HDF5 library: 1.14.2
-    netcdf library: 4.9.4-development
-    netCDF4: 1.7.2
-    h5netcdf: 1.3.0
-    h5py: 3.12.1
-    zarr: 3.1.3
-    s3fs: 2024.12.0
-    scipy: 1.15.1
-    dask: 2025.5.1
-    distributed: 2025.5.1
-    cftime: 1.6.4.post1
-    cfunits: 3.3.7
-    cfdm: 1.12.2.0
-    esmpy/ESMF: 8.7.0
-    psutil: 6.1.1
-    matplotlib: 3.10.0
-    cfplot: 3.4.0
-    cf: 3.18.0
 
     """
     # Get cfdm env