+++ cf/mixin/fielddomainlist.py
@@ -336,10 +336,10 @@ class FieldDomainList:
         """Convert the list elements to an `xarray` Dataset.
 
         If the `cf_xarray` package (https://cf-xarray.readthedocs.io)
-        is installed then the `cf_xarray` accessors will be present on
-        the returned `xarray` objects (`xarray.DataArray.cf` and
-        `xarray.Dataset.cf`) that allow some extra interpretation of
-        CF attributes.
+        is installed then the `cf_xarray` accessors that allow some
+        interpretation of CF attributes will bxe present on the
+        returned `xarray` objects (`xarray.DataArray.cf` and
+        `xarray.Dataset.cf`, but not `xarray.DataTree`).
 
         Note that ``ds = fl.to_xarray()`` is identical to ``ds =
         cf.write(fl, fmt='XARRAY')``
@@ -350,8 +350,11 @@ class FieldDomainList:
 
         :Returns:
 
-            `xarray.Dataset`
-                The equivalent `xarray` Dataset.
+            `xarray.Dataset` or `xarray.DataTree`
+                The equivalent `xarray` dataset. If there are no
+                sub-groups of the root group then an `xarray.Dataset`
+                is returned, oterwise an `xarray.DataTree` is
+                returned.
 
         """
         from cf.read_write import write