+++ cf/test/test_functions.py
@@ -55,7 +55,7 @@ class functionTest(unittest.TestCase):
         self.assertIsInstance(org, dict)
 
         # Check all keys that should be there are, with correct value type:
-        self.assertEqual(len(org), 12)  # update expected len if add new key(s)
+        self.assertEqual(len(org), 13)  # update expected len if add new key(s)
 
         # Types expected:
         self.assertIsInstance(org["atol"], float)
@@ -71,6 +71,7 @@ class functionTest(unittest.TestCase):
         # equiv. string
         self.assertIsInstance(org["log_level"], str)
         self.assertIsInstance(org["display_data"], bool)
+        self.assertIsInstance(org["persist_data"], bool)
 
         # Store some sensible values to reset items to for testing, ensuring:
         # 1) they are kept different to the defaults (i.e. org values); and