+++ cf/test/test_read_write.py
@@ -351,7 +351,13 @@ class read_writeTest(unittest.TestCase):
                 if ex_field_n in (8, 9, 10):
                     continue
 
-                cf.write(ex_field, tmpfile, fmt=fmt, mode="a")
+                cf.write(
+                    ex_field,
+                    tmpfile,
+                    fmt=fmt,
+                    mode="a",
+                    netcdf_backend="netCDF4",
+                )
                 f = cf.read(tmpfile)
 
                 if ex_field_n == 5:  # another special case
@@ -432,15 +438,25 @@ class read_writeTest(unittest.TestCase):
             # field n=5 which aggregates to one with n=2] => + 1 - 1 = + 0:
             overall_length = len(append_ex_fields)
             cf.write(
-                append_ex_fields, tmpfile, fmt=fmt, mode="a"
+                append_ex_fields,
+                tmpfile,
+                fmt=fmt,
+                mode="a",
+                netcdf_backend="netCDF4",
             )  # 2. now append
             f = cf.read(tmpfile)
             self.assertEqual(len(f), overall_length)
 
             # Also test the mode="r+" alias for mode="a".
-            cf.write(g, tmpfile, fmt=fmt, mode="w")  # 1. overwrite to wipe
             cf.write(
-                append_ex_fields, tmpfile, fmt=fmt, mode="r+"
+                g, tmpfile, fmt=fmt, mode="w", netcdf_backend="netCDF4"
+            )  # 1. overwrite to wipe
+            cf.write(
+                append_ex_fields,
+                tmpfile,
+                fmt=fmt,
+                mode="r+",
+                netcdf_backend="netCDF4",
             )  # 2. now append
             f = cf.read(tmpfile)
             self.assertEqual(len(f), overall_length)
@@ -542,8 +558,12 @@ class read_writeTest(unittest.TestCase):
             #     )
 
             # Check behaviour when append identical fields, as an edge case:
-            cf.write(g, tmpfile, fmt=fmt, mode="w")  # 1. overwrite to wipe
-            cf.write(g_copy, tmpfile, fmt=fmt, mode="a")  # 2. now append
+            cf.write(
+                g, tmpfile, fmt=fmt, mode="w", netcdf_backend="netCDF4"
+            )  # 1. overwrite to wipe
+            cf.write(
+                g_copy, tmpfile, fmt=fmt, mode="a", netcdf_backend="netCDF4"
+            )  # 2. now append
             f = cf.read(tmpfile)
             self.assertEqual(len(f), 2)
             self.assertTrue(
@@ -618,23 +638,29 @@ class read_writeTest(unittest.TestCase):
             )
 
     def test_read_write_unlimited(self):
-        for fmt in ("NETCDF4", "NETCDF3_CLASSIC"):
-            f = self.f1.copy()
+        f = cf.read(self.filename)[0]
+        for fmt in self.netcdf_fmts:
             domain_axes = f.domain_axes()
 
             domain_axes["domainaxis0"].nc_set_unlimited(True)
-            cf.write(f, tmpfile, fmt=fmt)
+            cf.write(f, tmpfile, fmt=fmt, cfa=None)
 
-            f = cf.read(tmpfile)[0]
-            domain_axes = f.domain_axes()
-            self.assertTrue(domain_axes["domainaxis0"].nc_is_unlimited())
+            if fmt in self.netcdf3_fmts:
+                # Note: netcdf_file backend does not support unlimited
+                #       dimensions
+                backend = "netCDF4"
+            else:
+                backend = None
 
-        fmt = "NETCDF4"
-        f = self.f1.copy()
+            g = cf.read(tmpfile, netcdf_backend=backend)[0]
+            domain_axes = g.domain_axes()
+            self.assertTrue(domain_axes["domainaxis0"].nc_is_unlimited(), fmt)
+
         domain_axes = f.domain_axes()
+
         domain_axes["domainaxis0"].nc_set_unlimited(True)
         domain_axes["domainaxis2"].nc_set_unlimited(True)
-        cf.write(f, tmpfile, fmt=fmt)
+        cf.write(f, tmpfile, fmt="NETCDF4", cfa=None)
 
         f = cf.read(tmpfile)[0]
         domain_axes = f.domain_axes()
@@ -842,7 +868,7 @@ class read_writeTest(unittest.TestCase):
         f = self.f1
         cf.write(f, tmpfile)
 
-        cf.write(f, tmpfile, omit_data="all")
+        cf.write(f, tmpfile, omit_data="all", netcdf_backend="netCDF4")
         g = cf.read(tmpfile)
         self.assertEqual(len(g), 1)
         g = g[0]
@@ -854,7 +880,12 @@ class read_writeTest(unittest.TestCase):
         # Check that a dump works
         g.dump(display=False)
 
-        cf.write(f, tmpfile, omit_data=("field", "dimension_coordinate"))
+        cf.write(
+            f,
+            tmpfile,
+            omit_data=("field", "dimension_coordinate"),
+            netcdf_backend="netCDF4",
+        )
         g = cf.read(tmpfile)[0]
 
         # Check that only the field and dimension coordinate data are
@@ -863,7 +894,7 @@ class read_writeTest(unittest.TestCase):
         self.assertFalse(np.ma.count(g.construct("grid_latitude").array))
         self.assertTrue(np.ma.count(g.construct("latitude").array))
 
-        cf.write(f, tmpfile, omit_data="field")
+        cf.write(f, tmpfile, omit_data="field", netcdf_backend="netCDF4")
         g = cf.read(tmpfile)[0]
 
         # Check that only the field data are missing
@@ -874,11 +905,10 @@ class read_writeTest(unittest.TestCase):
     #        True, "URL TEST: UNRELIABLE FLAKEY URL DESTINATION. TODO REPLACE URL"
     #    )
     def test_read_url(self):
-        """Test reading urls."""
+        """Test reading remote url."""
         for scheme in ("http", "https"):
             remote = f"{scheme}:///psl.noaa.gov/thredds/dodsC/Datasets/cru/crutem5/Monthlies/air.mon.anom.nobs.nc"
-            # Check that cf can access it
-            f = cf.read(remote)
+            f = cf.read(remote, netcdf_backend="netCDF4")
             self.assertEqual(len(f), 1)
 
     @unittest.skipUnless(
@@ -949,6 +979,82 @@ class read_writeTest(unittest.TestCase):
 
             z = cf.read(zarr_dataset, dataset_type="Zarr")
             self.assertEqual(len(z), 1)
+
+    def test_write_netcdf_backend(self):
+        """Test cf.write with different netCDF backends."""
+        f = self.f0
+
+        cf.write(f, tmpfile0, netcdf_backend="h5netcdf-h5py")
+        cf.write(f, tmpfile1, netcdf_backend="netCDF4")
+        f0 = cf.read(tmpfile0)[0]
+        f1 = cf.read(tmpfile1)[0]
+        self.assertTrue(f1.equals(f0))
+
+        f = cf.read(filename)
+        cf.write(f, tmpfile0, netcdf_backend="h5netcdf-h5py")
+        cf.write(f, tmpfile1, netcdf_backend="netCDF4")
+        f0 = cf.read(tmpfile0)[0]
+        f1 = cf.read(tmpfile1)[0]
+        self.assertTrue(f1.equals(f0))
+
+        # Bad fmt/backend combinations
+        for backend in ("netCDF4", "h5netcdf-h5py"):
+            with self.assertRaises(ValueError):
+                cf.write(f, tmpfile, fmt="ZARR3", netcdf_backend=backend)
+
+        for backend in ("zarr", "h5netcdf-h5py"):
+            with self.assertRaises(ValueError):
+                cf.write(
+                    f, tmpfile, fmt="NETCDF3_CLASSIC", netcdf_backend=backend
+                )
+
+        for backend in ("zarr",):
+            with self.assertRaises(ValueError):
+                cf.write(f, tmpfile, fmt="NETCDF4", netcdf_backend=backend)
+
+    def test_write_h5py_options(self):
+        """Test cf.write with h5py_options."""
+        f = self.f0
+        h5py_options = dict(
+            fs_strategy="page", fs_page_size=2**20, meta_block_size=500000
+        )
+
+        cf.write(f, tmpfile0, h5py_options=None)
+        size = os.path.getsize(tmpfile0)
+
+        cf.write(
+            f,
+            tmpfile1,
+            netcdf_backend="h5netcdf-h5py",
+            h5py_options=h5py_options,
+        )
+        self.assertTrue(os.path.getsize(tmpfile1) > size)
+
+        f0 = cf.read(tmpfile0)[0]
+        f1 = cf.read(tmpfile1)[0]
+        self.assertTrue(f1.equals(f0))
+
+        with self.assertRaises(ValueError):
+            cf.write(
+                f,
+                tmpfile0,
+                netcdf_backend="netCDF4",
+                h5py_options=h5py_options,
+            )
+
+        with self.assertRaises(ValueError):
+            cf.write(
+                f, tmpfile0, fmt="NETCDF3_CLASSIC", h5py_options=h5py_options
+            )
+
+    def test_read_netcdf_file(self):
+        """Test cf.read for differing the netcdf_file backend."""
+        f = self.f0
+
+        cf.write(f, tmpfile, fmt="NETCDF3_CLASSIC")
+        g = cf.read(tmpfile, netcdf_backend="netcdf_file")[0]
+
+        self.assertTrue(g.equals(f))
 
 
 if __name__ == "__main__":